Allow removing of plugins

+++ plugin-loader/src/pluginloader/Loader.py
@@ -16,54 +16,42 @@ import zipfile
 def loadGPG():
 	return gnupg.GPG(keyring='%s/plugins.keyring' % Board.pluginPath())
 
-class Loader(Plugin):
-	def __init__(self):
-		self.initializeKeychain()
-		self.loadPlugins()
-
-	def initializeKeychain(self):
-		filename = pkg_resources.resource_filename('pluginloader', 'files/telldus.gpg')
-		gpg = loadGPG()
-		installedKeys = [key['keyid'] for key in gpg.list_keys()]
-		defaultKeys = gpg.scan_keys(filename)
-		for key in [key['keyid'] for key in defaultKeys]:
-			if key in installedKeys:
-				continue
-			gpg.import_keys(open(filename, 'rb').read())
-			return
- 
-	def loadPlugins(self):
-		for f in glob.glob('%s/**/manifest.yml' % Board.pluginPath()):
-			try:
-				self.verifyAndLoadPlugin(f)
-			except Exception as e:
-				logging.error('Could not load plugin: %s', e)
+class LoadedPlugin(object):
+	def __init__(self, manifest, context):
+		self.loaded = False
+		self.context = context
+		self.verified = False
+		cfg = yaml.load(open(manifest, 'r').read())
+		self.name = cfg['name']
+		self.path = os.path.dirname(manifest)
+		self.packages = cfg['packages'] if 'packages' in cfg else []
 
 	def printBacktrace(self, bt):
 		for f in bt:
 			logging.error(str(f))
 
-	def verifyAndLoadPlugin(self, path):
-		cfg = yaml.load(open(path, 'r').read())
-		if 'packages' not in cfg:
-			return
-		d = os.path.dirname(path)
-		for p in cfg['packages']:
-			if not self.verifyPlugin('%s/%s' % (d, p)):
-				return
-		for p in cfg['packages']:
-			self.__loadPlugin('%s/%s' % (d, p))
+	def remove(self):
+		shutil.rmtree(self.path)
 
-	def verifyPlugin(self, path):
-		# Check signature
-		gpg = loadGPG()
-		v = gpg.verify_file(open('%s.asc' % path, 'rb'), path)
-		if v.valid is not True:
-			raise Exception('Could not verify signature: %s' % v.status)
+	def verify(self):
+		for p in self.packages:
+			if not LoadedPlugin.__verifyFile('%s/%s' % (self.path, p)):
+				return False
 		return True
 
-	def __loadPlugin(self, path):
-		for dist in pkg_resources.find_distributions(path):
+	def verifyAndLoad(self):
+		try:
+			if not self.verify():
+				return
+		except Exception as e:
+			logging.warning("Could not load plugin %s: %s", self.name, str(e))
+			return
+		for p in self.packages:
+			self.__loadEgg(p)
+		self.loaded = True
+
+	def __loadEgg(self, egg):
+		for dist in pkg_resources.find_distributions('%s/%s' % (self.path, egg)):
 			logging.warning("Loading plugin %s %s", dist.project_name, dist.version)
 			dist.activate()
 			for entry in dist.get_entry_map(group='telldus.plugins'):
@@ -90,6 +78,49 @@ class Loader(Plugin):
 					logging.error(str(e))
 					self.printBacktrace(traceback.extract_tb(exc_traceback))
 
+	@staticmethod
+	def __verifyFile(path):
+		# Check signature
+		gpg = loadGPG()
+		v = gpg.verify_file(open('%s.asc' % path, 'rb'), path)
+		if v.valid is not True:
+			raise Exception('Could not verify signature: %s' % v.status)
+		return True
+
+class Loader(Plugin):
+	def __init__(self):
+		self.plugins = []
+		self.initializeKeychain()
+		self.loadPlugins()
+
+	def initializeKeychain(self):
+		filename = pkg_resources.resource_filename('pluginloader', 'files/telldus.gpg')
+		gpg = loadGPG()
+		installedKeys = [key['keyid'] for key in gpg.list_keys()]
+		defaultKeys = gpg.scan_keys(filename)
+		for key in [key['keyid'] for key in defaultKeys]:
+			if key in installedKeys:
+				continue
+			gpg.import_keys(open(filename, 'rb').read())
+			return
+
+	def loadPlugin(self, manifest):
+		plugin = LoadedPlugin(manifest, self.context)
+		plugin.verifyAndLoad()
+		self.plugins.append(plugin)
+
+	def loadPlugins(self):
+		for f in glob.glob('%s/**/manifest.yml' % Board.pluginPath()):
+			self.loadPlugin(f)
+
+	def removePlugin(self, name):
+		for i, plugin in enumerate(self.plugins):
+			if plugin.name != name:
+				continue
+			plugin.remove()
+			del self.plugins[i]
+			return
+
 class WebFrontend(Plugin):
 	implements(IWebRequestHandler)
 
@@ -115,7 +146,7 @@ class WebFrontend(Plugin):
 				os.unlink(k)
 		except Exception as e:
 			os.unlink(filename)
-			return 'pluginloader.html', {'msg': str(e)}
+			return 'pluginloader.html', {'msg': str(e), 'loader': Loader(self.context)}
 		return WebResponseRedirect('/')
 
 	def importPlugin(self):
@@ -156,9 +187,11 @@ class WebFrontend(Plugin):
 			for p, s in packages:
 				os.rename(p, '%s/%s' % (path, os.path.basename(p)))
 				z.extract(s, path)
-			z.extract(info, path)
+			manifest = z.extract(info, path)
 		os.unlink(filename)
-		return 'pluginloader.html', {'msg': 'Plugin was imported'}
+		loader = Loader(self.context)
+		loader.loadPlugin(manifest)
+		return 'pluginloader.html', {'msg': 'Plugin was imported', 'loader': loader}
 
 	def uploadPlugin(self, f):
 		with open('%s/staging.zip' % (Board.pluginPath()), 'w') as wf:
@@ -170,7 +203,7 @@ class WebFrontend(Plugin):
 	def matchRequest(self, plugin, path):
 		if plugin != 'pluginloader':
 			return False
-		if path in ['', 'importkey']:
+		if path in ['', 'importkey', 'remove']:
 			return True
 		return False
 
@@ -178,15 +211,22 @@ class WebFrontend(Plugin):
 		if request.method() == 'POST' and 'pluginfile' in params:
 			self.uploadPlugin(params['pluginfile'])
 			return WebResponseRedirect('/')
+
 		if path == 'importkey':
 			if 'discard' in params:
 				os.unlink('%s/staging.zip' % (Board.pluginPath()))
 				return WebResponseRedirect('/')
 			return self.importKey(params['key'] if 'key' in params else None)
+
+		if path == 'remove':
+			if 'pluginname' in params:
+				Loader(self.context).removePlugin(params['pluginname'])
+			return WebResponseRedirect('/')
+
 		if os.path.isfile('%s/staging.zip' % (Board.pluginPath())):
 			try:
 				return self.importPlugin()
 			except ImportError as e:
 				os.unlink('%s/staging.zip' % (Board.pluginPath()))
-				return 'pluginloader.html', {'msg':'Error importing plugin: %s' % e}
-		return 'pluginloader.html', {'msg':''}
+				return 'pluginloader.html', {'msg':'Error importing plugin: %s' % e, 'loader': Loader(self.context)}
+		return 'pluginloader.html', {'msg':'', 'loader': Loader(self.context)}
